Use mailgun to send notifications
With https://github.com/SuffolkLITLab/ALActions/pull/15 merged, we can use
Mailgun to send error notification emails now!

There are a few secrets that need to be set in GitHub for this to work:

* MAILGUN_API_KEY: can be the domain sending key, or the full API key
* MAILGUN_DOMAIN
* ERROR_EMAILS: a comma separated list of the emails we want to get
  notifications for something going wrong with the servers.
* ERROR_EMAIL_FROM: the email with want to send the message from. Can
  be the name convention, "John Brown <example@example.com>", so the email
  is easier to recoginize. Also important to make sure that the message
  is sent from an authorized sender so it doesn't get filtered a spam
  (which kept happening with my gmail).

[GitHub's documenation for secrets](https://docs.github.com/en/actions/security-guides/using-secrets-in-github-actions)
will help when setting those up. Since we have a repository for maintenance
stuff, I recommend making the repo secrets. We can always make them
organization secrets later, but better to keep things scoped for now.

Happy to hop on a call to discuss more or help out setting up at all.